provisioning-solution (Python)/provisioning_requirements.txt
azure-functions
msal
requests
azure-servicebus
azure-storage-blob
python-dateutil
